ARM(ALU): Moved ALU tests to `data_processing.rs`
